+++ src/php/Api/Infrastructure/PhelFnLoaderInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phel\Api\Infrastructure;
+
+use Phel\Lang\Collections\Map\PersistentMapInterface;
+
+interface PhelFnLoaderInterface
+{
+    /**
+     * @param list<string> $namespaces
+     *
+     * @return array<string,PersistentMapInterface>
+     */
+    public function getNormalizedPhelFunctions(array $namespaces = []): array;
+}